(search) Divide redux store to modules
