--- applications/search/src/redux/ActionTypes.js
@@ -1,31 +1,0 @@
-export const DATASETS_REQUEST = 'DATASETS_REQUEST';
-export const DATASETS_SUCCESS = 'DATASETS_SUCCESS';
-export const DATASETS_FAILURE = 'DATASETS_FAILURE';
-
-export const DATASETDETAILS_REQUEST = 'DATASETDETAILS_REQUEST';
-export const DATASETDETAILS_SUCCESS = 'DATASETDETAILS_SUCCESS';
-export const DATASETDETAILS_FAILURE = 'DATASETDETAILS_FAILURE';
-
-export const TERMS_REQUEST = 'TERMS_REQUEST';
-export const TERMS_SUCCESS = 'TERMS_SUCCESS';
-export const TERMS_FAILURE = 'TERMS_FAILURE';
-
-export const THEMES_REQUEST = 'THEMES_REQUEST';
-export const THEMES_SUCCESS = 'THEMES_SUCCESS';
-export const THEMES_FAILURE = 'THEMES_FAILURE';
-
-export const PUBLISHERS_REQUEST = 'PUBLISHERS_REQUEST';
-export const PUBLISHERS_SUCCESS = 'PUBLISHERS_SUCCESS';
-export const PUBLISHERS_FAILURE = 'PUBLISHERS_FAILURE';
-
-export const OPENLICENSES_REQUEST = 'OPENLICENSES_REQUEST';
-export const OPENLICENSES_SUCCESS = 'OPENLICENSES_SUCCESS';
-export const OPENLICENSES_FAILURE = 'OPENLICENSES_FAILURE';
-
-export const DISTRIBUTIONTYPE_REQUEST = 'DISTRIBUTIONTYPE_REQUEST';
-export const DISTRIBUTIONTYPE_SUCCESS = 'DISTRIBUTIONTYPE_SUCCESS';
-export const DISTRIBUTIONTYPE_FAILURE = 'DISTRIBUTIONTYPE_FAILURE';
-
-export const DATASETDETAILS_RESET = 'DATASETDETAILS_RESET';
-
-export const SETTINGS_PATCH = 'SETTINGS_PATCH';
