+++ applications/search/src/redux/modules/featureToggle.js
